Fix Go duplicate TestExists compilation failure
Remove exists_test.go from Go template directory and generate it
programmatically in Test_go.ts. This prevents a duplicate TestExists
declaration when a scaffolded projectname_sdk_test.go already exists
in the test directory, which caused `go test ./...` to fail.
