--- project/.sdk/tm/go/test/exists_test.go
@@ -1,16 +1,0 @@
-package sdktest
-
-import (
-	"testing"
-
-	sdk "GOMODULE"
-)
-
-func TestExists(t *testing.T) {
-	t.Run("test-mode", func(t *testing.T) {
-		testsdk := sdk.TestSDK(nil, nil)
-		if testsdk == nil {
-			t.Fatal("expected non-nil SDK")
-		}
-	})
-}
